+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Hsu.
 * Docker commands are now done directly via Docker API instead of shelling out
 to Docker CLI.
 * [GH-27] Added `--data-bag-path` option. Copies over data bags.
+* [FSE-201] Update knife-container to support ChefDK and Chef 12 libraries
 
 ## v0.2.3 (2014-09-19)
 * [GH-39] Fixed `--dockerfiles-path` parameter which did not properly accept input.